now we can get the login URL.
I already did some error handling although I prefer keeping quiet and returning false, instead of giving out the details to everyone. a log will be taken in the txt file.

be sure to use === when verifying that the User has no 2FA or whether it's an error, although you can just use == and log the user in to be error tolerant.